Trainer img added, recurrence fixed, email validation fixed
