+++ Yoga/src/main/webapp/css/common.css
@@ -282,3 +282,17 @@ input[type=time] {
 .trainer-schedules .trainer-details-card-ineer-container {
 	background: #f3f8ff;
 }
+
+.bh-img {
+	overflow: hidden;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	background: white;
+	margin-top: 20px;
+	height: 256px;
+	width: 256px;
+	border-radius: 50%;
+	box-shadow: 0 2px 5px 0 rgb(0 0 0/ 16%), 0 2px 10px 0 rgb(0 0 0/ 12%);
+	border: 5px solid white;
+}
